removed unneeded casts

+++ Teleopti.Analytics/Teleopti.Analytics.Etl.TransformerInfrastructure/GeneralInfrastructure.cs
@@ -75,7 +75,7 @@ namespace Teleopti.Analytics.Etl.TransformerInfrastructure
 			{
 				foreach (DataRow dataRow in dataSet.Tables[0].Rows)
 				{
-					ITimeZoneDim timeZoneDim = new TimeZoneDim((int)(Int16)dataRow["time_zone_id"],
+					ITimeZoneDim timeZoneDim = new TimeZoneDim((Int16)dataRow["time_zone_id"],
 														(string)dataRow["time_zone_code"],
 														(string)dataRow["time_zone_name"],
 														handleDbNullValueBool(dataRow["default_zone"]),
